// src/types/lightspeed.ts
export interface PaginatedResponse<T> {
  data: T[];
  version?: {
    min: number;
    max: number;
  };
}

export interface SingleResponse<T> {
  data: T;
  includes?: unknown;
}
// ...
export interface Brand {
  id: string;
  name: string;
  description?: string;
  created_at?: string;
  updated_at?: string;
  deleted_at?: string;
  version?: number;
}

export interface BrandCreateInput {
  name: string;
  description?: string;
}

// Product Type types
export interface ProductType {
  id: string;
  name: string;
  created_at?: string;
  updated_at?: string;
  deleted_at?: string;
  version?: number;
}

export interface ProductTypeCreateInput {
  name: string;
}

// Tag types
export interface Tag {
  id: string;
  name: string;
  created_at?: string;
  updated_at?: string;
  deleted_at?: string;
  version?: number;
}

export interface TagCreateInput {
  name: string;
}

// Supplier types
export interface Supplier {
  id: string;
  name: string;
  description?: string;
  source?: string;
  contact?: SupplierContact;
  created_at?: string;
  updated_at?: string;
  deleted_at?: string;
}

export interface SupplierContact {
  first_name?: string;
  last_name?: string;
  company_name?: string;
  phone?: string;
  mobile?: string;
  fax?: string;
  email?: string;
  twitter?: string;
  website?: string;
  physical_address_1?: string;
  physical_address_2?: string;
  physical_suburb?: string;
  physical_city?: string;
  physical_postcode?: string;
  physical_state?: string;
  physical_country_id?: string;
  postal_address_1?: string;
  postal_address_2?: string;
  postal_suburb?: string;
  postal_city?: string;
  postal_postcode?: string;
  postal_state?: string;
  postal_country_id?: string;
}

export interface SupplierCreateInput {
  name: string;
  description?: string;
  contact?: SupplierContact;
}